solved some bugs

+++ app/src/main/java/com/example/p2pgeocaching/data/FeedData.kt
@@ -3,6 +3,7 @@ package com.example.p2pgeocaching.data
 import android.provider.SyncStateContract
 import android.util.Log
 import com.example.p2pgeocaching.activities.NewCacheActivity
+import com.example.p2pgeocaching.bluetooth.BluetoothHandler
 import com.example.p2pgeocaching.caches.CacheListGenerator
 import com.example.p2pgeocaching.constants.Constants
 import java.io.File
@@ -17,6 +18,16 @@ import com.example.p2pgeocaching.ownbacnet.OwnFeed
  */
 @kotlinx.serialization.Serializable
 class FeedData : Serializable {
+
+    var stringFileContent: String = ""
+
+    constructor(context: File) {
+        stringFileContent = feedToData(context)
+    }
+
+    constructor(file: File, context: File) {
+        dataToFeed(file, context)
+    }
 
     companion object {
         const val TAG = "FeedData"
@@ -38,11 +49,12 @@ class FeedData : Serializable {
         val pubkey = keys[0].split("_")
         val salt = pubkey[1].takeLast(4)
         val feedname = userName.plus("#").plus(salt)
-
+        Log.i(TAG, "feedName = "+feedname)
         val ownFeedFile = File(context, feedname)
         val ownFeedContent = ownFeedFile.readText()
-
+        Log.i(TAG, "Before if")
         if (ownFeedContent.length != 0) {
+            Log.i(TAG, "length is not zero")
             file.appendText(feedname)
             file.appendText("_:_:_")
             file.appendText("".plus(ownFeedContent))
@@ -50,28 +62,32 @@ class FeedData : Serializable {
         }
 
         var feedNamesFile = File(context, Constants.FEED_NAMES_FILE)
-        var feedNameList = feedNamesFile.readText().split("\n")
-        for (feed in feedNameList) {
-            val feedFile = File(context, feed)
-            val feedContent = feedFile.readText()
-            if (file.length() == 0L && feedContent.length != 0) {
-                file.appendText(feed)
-                file.appendText("_:_:_")
-                file.appendText("".plus(feedContent))
-                Log.i(TAG, "File Content after adding new feed = "+file.readText())
-            } else if (feedContent.length != 0) {
-                file.appendText("#####")
-                file.appendText("".plus(feed))
-                file.appendText("_:_:_")
-                file.appendText("".plus(feedContent))
-                Log.i(TAG, "File Content after adding new feed = "+file.readText())
+        if (feedNamesFile.length() != 0L) {
+            var feedNameList = feedNamesFile.readText().split("\n")
+
+            for (feed in feedNameList) {
+                val feedFile = File(context, feed)
+                val feedContent = feedFile.readText()
+                if (file.length() == 0L && feedContent.length != 0) {
+                    file.appendText(feed)
+                    file.appendText("_:_:_")
+                    file.appendText("".plus(feedContent))
+                    Log.i(TAG, "File Content after adding new feed = " + file.readText())
+                } else if (feedContent.length != 0) {
+                    file.appendText("#####")
+                    file.appendText("".plus(feed))
+                    file.appendText("_:_:_")
+                    file.appendText("".plus(feedContent))
+                    Log.i(TAG, "File Content after adding new feed = " + file.readText())
+                }
             }
         }
         Log.i(TAG, "File Content at the end = "+file.readText())
-        return "file"
+        return file.readText()
     }
 
     fun dataToFeed (file: File, context:File) {
+        Log.i(TAG, "started dataToFeed")
         val fileContent = file.readText()
         Log.i(TAG, "Received File Content = "+fileContent)
         val feedList = fileContent.split("#####")
@@ -94,29 +110,42 @@ class FeedData : Serializable {
                 if (person.equals(feedName)) {
                     Log.i(TAG, "Feed is subscribed")
                     val personFile = File(context, person)
-                    val entrylist = fdp.feedToEntrylist(personFile)
-                    var len = entrylist.size
-                    val lastEntry = entrylist[len-1]
-
-                    val supportFile = File(context, "support")
-                    supportFile.createNewFile()
-                    supportFile.appendText(feedContent)
-                    Log.i(TAG, "Support File Content = "+supportFile.readText())
-                    val feedEntryList = fdp.feedToEntrylist(supportFile)
-                    len = feedEntryList.size
-                    val lastEntryInFeed = feedEntryList[len-1]
-
-                    Log.i(TAG, "Old Feed Content = "+personFile.readText())
-
-                    if (lastEntry.id < lastEntryInFeed.id) {
-                        for (i in (lastEntry.id+1)..lastEntryInFeed.id) {
+                    if (personFile.length() == 0L) {
+                        personFile.appendText(feedContent)
+                        val feedEntryList = fdp.feedToEntrylist(personFile)
+                        for (i in 0..(feedEntryList.size-1)) {
                             newEntryList.add(feedEntryList[i])
                         }
-                        personFile.delete()
-                        personFile.createNewFile()
-                        personFile.appendText(supportFile.readText())
-                        Log.i(TAG, "New Feed Content = "+personFile.readText())
-                        break;
+                    } else {
+                        val entrylist = fdp.feedToEntrylist(personFile)
+                        var len = entrylist.size
+                        val lastEntry = entrylist[len - 1]
+
+                        val supportFile = File(context, "support")
+                        if (!supportFile.exists()) {
+                            supportFile.createNewFile()
+                        } else {
+                            supportFile.delete()
+                            supportFile.createNewFile()
+                        }
+                        supportFile.appendText(feedContent)
+                        Log.i(TAG, "Support File Content = " + supportFile.readText())
+                        val feedEntryList = fdp.feedToEntrylist(supportFile)
+                        len = feedEntryList.size
+                        val lastEntryInFeed = feedEntryList[len - 1]
+
+                        Log.i(TAG, "Old Feed Content = " + personFile.readText())
+
+                        if (lastEntry.id < lastEntryInFeed.id) {
+                            for (i in (lastEntry.id + 1)..lastEntryInFeed.id) {
+                                newEntryList.add(feedEntryList[i])
+                            }
+                            personFile.delete()
+                            personFile.createNewFile()
+                            personFile.appendText(supportFile.readText())
+                            Log.i(TAG, "New Feed Content = " + personFile.readText())
+                            break;
+                        }
                     }
                 }
             }
